Initial commit of ns-net-bridge
This tool allows for a process to "bridge the gap" between a docker
container listening only on localhost and the host machine.

This is very useful for something like node debugging, where node
will open a port that's scoped to localhost, but nothing outside
the container can access it.

With this tool you can. It's easier to use the `scripts/connect`
script to do the connection, it will take a running docker container
and call the real powerhouse with all the right arguments. If you need
something more complicated, checkout the help for ns-net-bridge.